\name{getWELLState}
\alias{getWELLState}
\title{Get the state of a WELL generator implemented in randtoolbox package.}

\description{
  Get the state of a WELL generator implemented in randtoolbox package.
}
  
\usage{ 
  getWELLState()
} 

\value{
  An 0,1-matrix, whose columns represent 32-bit integers.
}


