\name{getWELLState}
\alias{getWELLState}
\title{Get the state of a WELL generator implemented in randtoolbox package.}

\description{
  Get the state of a WELL generator implemented in randtoolbox package
  to be used in function \code{rngWELLScriptR()}.
}
  
\usage{ 
  getWELLState()
} 

\value{
  A 0,1-matrix, whose columns represent 32-bit integers.
}

\seealso{
  \code{\link{rngWELLScriptR}}
}

