\name{randtoolbox-package}
\alias{randtoolbox-package}
\alias{randtoolbox}
\docType{package}
\title{
  Toolbox for pseudo and quasi random number generation
}
\description{The randtoolbox-package started in 2007 
  during an ISFA (France) working group. From then, it
  grew quickly thanks to the contribution of Diethelm Wuertz
  and Petr Savicky. It was presented in the Rmetrics 
  Meielisalp workshop in 2009. Slides can be found
  here: \url{http://www.rmetrics.org/Meielisalp2009/Presentations/Dutang.pdf}.
   
  This package has currently implementations for 
  state-of-the-art pseudo RNGs for simulations as well as the usual quasi RNGs.
  There are also some RNG tests. See \code{?pseudo.randtoolbox}
  and \code{?quasi.randtoolbox}.

  We recommend first users to take a look at the first
  part of the vignette. The second part is for people who
  want a deeper presentation of this topic.
  
  The stable version is available on CRAN 
  \url{http://cran.r-project.org/web/packages/randtoolbox/index.html},
  while the development version is hosted on R-forge
  \url{http://r-forge.r-project.org/projects/rmetrics/}.
  
} 
\details{
\tabular{ll}{
Package: \tab randtoolbox\cr
Type: \tab Package\cr
Version: \tab 1.09\cr
Date: \tab 2009-10-30\cr
License: \tab GPL version 2 or later\cr
}
}
\author{
Christophe Dutang and Petr Savicky
}
\keyword{ package }


