\name{runifinterface}
\alias{runifinterface}
\alias{set.generator}
\alias{put.state}
\alias{get.state}
\title{TODO}
\description{
TODO
}
\usage{
set.generator(name=c("congruRand", "default"), parameters=NULL, seed=NULL, ...,	only.description=FALSE)
put.state(description)		
get.state()
}
\arguments{
	\item{name}{}
	\item{parameters}{}	
	\item{seed}{}	
	\item{...}{}	
	\item{only.description}{}	
	\item{description}{}	
}
\details{
TODO
}
\value{
TODO
}
\seealso{
TODO
}
\author{
 Petr Savicky
}
\examples{

\dontrun{
TODO
}

}
\keyword{distribution}

